move env main file contents to common
