+++ scopes/react/react/templates/react-env/files/extension.ts
@@ -1,25 +1,6 @@
 import { ComponentContext } from '@teambit/generator';
+import { MainFileContents } from '../../common/common-env-main';
 
-export function extensionFile({ namePascalCase: Name }: ComponentContext) {
-  return `import { EnvsMain, EnvsAspect } from '@teambit/envs'
-import { ReactAspect, ReactMain } from '@teambit/react'
-
-export class ${Name}Extension {
-  constructor(private react: ReactMain) {}
-
-  static dependencies: any = [EnvsAspect, ReactAspect]
-
-  static async provider([envs, react]: [EnvsMain, ReactMain]) {
-    const ${Name}Env = react.compose([
-      /*
-        Use any of the "react.override..." transformers to
-      */
-    ])
-
-    envs.registerEnv(${Name}Env)
-
-    return new ${Name}Extension(react)
-  }
-}
-`;
+export function extensionFile(context: ComponentContext) {
+  return MainFileContents(context);
 }
